+++ lib/view_component/collection.rb
@@ -4,14 +4,17 @@ require "action_view/renderer/collection_renderer" if Rails.version.to_f >= 6.1
 
 module ViewComponent
   class Collection
+    attr_reader :component
+    delegate :format, to: :component
+
     def render_in(view_context, &block)
       iterator = ActionView::PartialIteration.new(@collection.size)
 
-      @component.compile(raise_errors: true)
-      @component.validate_collection_parameter!(validate_default: true)
+      component.compile(raise_errors: true)
+      component.validate_collection_parameter!(validate_default: true)
 
       @collection.map do |item|
-        content = @component.new(**component_options(item, iterator)).render_in(view_context, &block)
+        content = component.new(**component_options(item, iterator)).render_in(view_context, &block)
         iterator.iterate!
         content
       end.join.html_safe
@@ -34,8 +37,8 @@ module ViewComponent
     end
 
     def component_options(item, iterator)
-      item_options = { @component.collection_parameter => item }
-      item_options[@component.collection_counter_parameter] = iterator.index + 1 if @component.counter_argument_present?
+      item_options = { component.collection_parameter => item }
+      item_options[component.collection_counter_parameter] = iterator.index + 1 if component.counter_argument_present?
 
       @options.merge(item_options)
     end