+++ test/view_component/integration_test.rb
@@ -330,6 +330,17 @@ class IntegrationTest < ActionDispatch::IntegrationTest
     assert_select("p", text: "Mints counter: 2")
   end
 
+  def test_renders_inline_collections
+    get "/inline_products"
+
+    assert_select("h1", text: "Product", count: 2)
+    assert_select("h2", text: "Radio clock")
+    assert_select("h2", text: "Mints")
+    assert_select("p", text: "Today only", count: 2)
+    assert_select("p", text: "Radio clock counter: 1")
+    assert_select("p", text: "Mints counter: 2")
+  end
+
   def test_renders_the_previews_in_the_configured_route
     with_preview_route("/previews") do
       get "/previews"